# openapi-schema.yml
openapi: 3.0.3
info:
  title: SaaS Project Tracker API
  version: 1.0.0
  description: Backend API for the SaaS Project Tracker.
paths:
...
  /api/v1/issues/{issue_id}/:
    get:
      operationId: v1_issues_retrieve
      parameters:
      - in: path
        name: issue_id
        schema:
          type: integer
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
          description: ''
    patch:
      operationId: v1_issues_partial_update
      parameters:
      - in: path
        name: issue_id
        schema:
          type: integer
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedIssue'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedIssue'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedIssue'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
          description: ''
  /api/v1/issues/{issue_id}/activity/:
    get:
      operationId: issue_activity_list
      description: |-
        Read-only history of one issue. Access follows the parent project's
        membership, so non-members get the same 404 as for a missing issue.
      parameters:
      - in: query
        name: action
        schema:
          type: string
          enum:
          - ISSUE_ARCHIVED
          - ISSUE_ASSIGNED
          - ISSUE_CREATED
          - ISSUE_STATUS_CHANGED
          - ISSUE_UNASSIGNED
          - ISSUE_UPDATED
          - MEMBER_ADDED
          - MEMBER_REMOVED
          - PROJECT_ARCHIVED
          - PROJECT_CREATED
          - PROJECT_UPDATED
          - ROLE_CHANGED
      - in: path
        name: issue_id
        schema:
          type: integer
        required: true
      - in: query
        name: page
        schema:
          type: integer
      - in: query
        name: page_size
        schema:
          type: integer
      - in: query
        name: user
        schema:
          type: integer
        description: Id of the user who performed the action.
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedActivity'
          description: ''
...
  /api/v1/issues/{issue_id}/transition/:
    post:
      operationId: v1_issues_transition_create
      parameters:
      - in: path
        name: issue_id
        schema:
          type: integer
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueTransition'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/IssueTransition'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/IssueTransition'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
          description: ''
...
  /api/v1/projects/{project_id}/:
    get:
      operationId: project_retrieve
      parameters:
      - in: path
        name: project_id
        schema:
          type: integer
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
    patch:
      operationId: v1_projects_partial_update
      parameters:
      - in: path
        name: project_id
        schema:
          type: integer
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedProject'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedProject'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedProject'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
  /api/v1/projects/{project_id}/activity/:
    get:
      operationId: project_activity_list
      description: |-
        Read-only history of a project. Any current member may read it,
        including viewers, members who joined later, and archived projects.
      parameters:
      - in: query
        name: action
        schema:
          type: string
          enum:
          - ISSUE_ARCHIVED
          - ISSUE_ASSIGNED
          - ISSUE_CREATED
          - ISSUE_STATUS_CHANGED
          - ISSUE_UNASSIGNED
          - ISSUE_UPDATED
          - MEMBER_ADDED
          - MEMBER_REMOVED
          - PROJECT_ARCHIVED
          - PROJECT_CREATED
          - PROJECT_UPDATED
          - ROLE_CHANGED
      - in: query
        name: entity_id
        schema:
          type: integer
        description: Requires entity_type.
      - in: query
        name: entity_type
        schema:
          type: string
          enum:
          - ISSUE
          - MEMBERSHIP
          - PROJECT
      - in: query
        name: page
        schema:
          type: integer
      - in: query
        name: page_size
        schema:
          type: integer
      - in: path
        name: project_id
        schema:
          type: integer
        required: true
      - in: query
        name: user
        schema:
          type: integer
        description: Id of the user who performed the action.
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedActivity'
          description: ''
...
  /api/v1/projects/{project_id}/issues/:
    get:
      operationId: v1_projects_issues_retrieve
      parameters:
      - in: query
        name: assignee
        schema:
          type: integer
      - in: query
        name: issue_type
        schema:
          type: string
      - in: query
        name: page
        schema:
          type: integer
      - in: query
        name: page_size
        schema:
          type: integer
      - in: query
        name: priority
        schema:
          type: string
      - in: path
        name: project_id
        schema:
          type: integer
        required: true
      - in: query
        name: status
        schema:
          type: string
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedIssue'
          description: ''
    post:
      operationId: v1_projects_issues_create
      parameters:
      - in: path
        name: project_id
        schema:
          type: integer
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Issue'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Issue'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Issue'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
          description: ''
  /api/v1/projects/{project_id}/members/:
    get:
      operationId: v1_projects_members_list
      parameters:
      - in: path
        name: project_id
        schema:
          type: integer
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectMember'
          description: ''
    post:
      operationId: v1_projects_members_create
      parameters:
      - in: path
        name: project_id
        schema:
          type: integer
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProjectMember'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AddProjectMember'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AddProjectMember'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMember'
          description: ''
...
components:
  schemas:
...
    AddProjectMember:
      type: object
      properties:
        user_id:
          type: integer
        role:
          allOf:
          - $ref: '#/components/schemas/ProjectRoleEnum'
          default: TEAM_MEMBER
      required:
      - user_id
...
    Issue:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        issue_key:
          type: string
          readOnly: true
        issue_number:
          type: integer
          readOnly: true
        project:
          type: integer
          readOnly: true
        title:
          type: string
          maxLength: 255
        description:
          type: string
        issue_type:
          $ref: '#/components/schemas/IssueTypeEnum'
        priority:
          $ref: '#/components/schemas/PriorityEnum'
        status:
          allOf:
          - $ref: '#/components/schemas/StatusEnum'
          readOnly: true
        reporter:
          type: string
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          readOnly: true
        assignee:
          type: integer
          nullable: true
        reporter_detail:
          type: object
          nullable: true
          properties:
            id:
              type: integer
            username:
              type: string
          readOnly: true
        assignee_detail:
          type: object
          nullable: true
          properties:
            id:
              type: integer
            username:
              type: string
          readOnly: true
        due_date:
          type: string
          format: date
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        is_archived:
          type: boolean
          readOnly: true
      required:
      - assignee_detail
      - created_at
      - id
      - is_archived
      - issue_key
      - issue_number
      - project
      - reporter
      - reporter_detail
      - status
      - title
      - updated_at
    IssueTransition:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/StatusEnum'
      required:
      - status
...
    PaginatedActivity:
      type: object
      properties:
        count:
          type: integer
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true
        results:
          type: array
          items:
            $ref: '#/components/schemas/Activity'
      required:
      - count
      - next
      - previous
      - results
    PaginatedIssue:
      type: object
      properties:
        count:
          type: integer
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true
        results:
          type: array
          items:
            $ref: '#/components/schemas/Issue'
      required:
      - count
      - next
      - previous
      - results
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT